prishepka/requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1  
Django              3.2.5  
django-cors-headers 3.7.0  
djangorestframework 3.12.4 
gunicorn            20.1.0 
Pillow              8.3.1  
pip                 20.0.2 
pkg-resources       0.0.0  
psycopg2-binary     2.9.1  
pytz                2021.1 
setuptools          44.0.0 
sqlparse            0.4.1  
